Rename createFeed.ts to createFeedHelper.ts
Import statements on Mac are case-insensitive, causing collision with CreateFeed.tsx
