+++ src/components/CreateFeed/CreateFeed.tsx
@@ -21,7 +21,7 @@ import PipelineContainer from "./PipelineContainter";
 import Review from "./Review";
 import withSelectionAlert from "./SelectionAlert";
 import { useTypedSelector } from "../../store/hooks";
-import { createFeed } from "./createFeed";
+import { createFeed } from "./createFeedHelper";
 import "./createFeed.css";
 
 export default function CreateFeed() {
